Fixing bbone_chain bug
bone chain was dependent on when they were added to the armature not on its bone hierarchy. This was fix by making sure the parent bone was the first on a list and then all the recursive children in order.
Also, there were minors tweaks on single_control.py, sorting the functions alphabetically so it easier to look up in case of changes.